deep-researcher-backend-beta/pyproject.toml
[project]
name = "deep-researcher-backend-beta"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "crawl4ai>=0.7.4",
    "extract-favicon>=0.5.4",
    "fastapi>=0.118.0",
    "google-genai>=1.42.0",
    "googlesearch-python>=1.3.0",
    "markdown-pdf>=1.10",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "pillow>=11.3.0",
    "py-youtube>=1.1.7",
    "pypandoc>=1.15",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.37.0",
]
